Comprehensive security hardening, Decimal migration, and new features
Security & Integrity:
- Fix QuickBooks OAuth CSRF with state parameter validation
- Make journal entry creation atomic with balance validation
- Add validate_number on all monetary schema fields
- Add date format validation across all date fields
- Fix EmailDigestWorker crash on deleted users
- Document single-tenant design assumption in README

Float to Decimal Migration:
- Create Holdco.Money module with nil-safe Decimal arithmetic helpers
- Convert all ~60 monetary :float fields to :decimal across 35 schemas
- Update all business logic (Portfolio, Finance, Analytics, Depreciation,
  CapitalGains, Scenarios) to use Money module
- Fix all LiveView files (~25 files) to use Decimal-safe operations
- Fix controllers (export, report_html, xbrl) for Decimal compatibility
- Update all test assertions to handle Decimal comparisons

New Features:
- Contacts: interaction history tracking with notes and types
- Projects: milestone tracking with status and due dates
- Approvals: N-of-M voting logic with approval_votes table
- Webhooks: event subscription UI in settings
- KPIs: auto-population from financial data with data_source field
- Per-company QuickBooks integration (company-scoped connections)
- Global AI chat drawer (sticky across all authenticated pages)

+++ lib/holdco/assets/asset_holding.ex
@@ -5,7 +5,7 @@ defmodule Holdco.Assets.AssetHolding do
   schema "asset_holdings" do
     field :asset, :string
     field :ticker, :string
-    field :quantity, :float
+    field :quantity, :decimal
     field :unit, :string
     field :currency, :string, default: "USD"
     field :asset_type, :string, default: "other"
